requirements.txt
Brotli
brotlicffi
certifi
cffi
charset-normalizer
contourpy
cycler
filelock
fonttools
gmpy2
idna
imageio==2.37.0
importlib_resources
Jinja2
joblib
kiwisolver
lazy_loader==0.4
llvmlite
MarkupSafe
matplotlib==3.9.1
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
mpmath
networkx
numba
numpy
opencv-python==4.7.0
packaging
pillow
ply==3.11
pycparser
pyparsing
PyQt5==5.15.9
PyQt5-sip==12.12.2
PySocks
python-dateutil
PyYAML
requests
scikit-image==0.24.0
scikit-learn
scipy
sip
six
sympy
threadpoolctl
tifffile==2024.8.30
toml
tomli
torch==2.5.1
torchvision
tornado
triton==3.1.0
typing_extensions
unicodedata2
urllib3
zipp
click
Flask
importlib_metadata
itsdangerous
Werkzeug